pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
name = "nicheformer"
version = "0.0.1"
description = "Nicheformer: a foundation model for single-cell and spatial omics"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Anna Schaar"},
    {name = "Alejandro Tejada-Lapuerta"},
]
maintainers = [
    {name = "Anna Schaar", email = "anna.schaar@tum.de"},
    {name = "Alejandro Tejada-Lapuerta", email = "alejandro.tejada@helmholtz-munich.de"},
]
urls.Documentation = "https://nicheformer.readthedocs.io/"
urls.Source = "https://github.com/theislab/nicheformer"
urls.Home-page = "https://github.com/theislab/nicheformer"
dependencies = [
    "anndata",
    "session-info",
    "cellxgene-schema==3.1.3",
    "squidpy",
    "pyensembl==2.3.13",
    "torch>=2.5.1",
    "pytorch-lightning>=2.0.0",
    "numpy==1.26.4",
    "scikit-learn>=1.2.0",
    "scipy>=1.10.0",
    "pandas==1.5.3",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "merlin-core>=23.8.0",
    "merlin-dataloader>=23.8.0",
    "dask>=2024.1.1",
    "dask-cuda>=24.2.0",
    "torchmetrics>=1.4.0",
    "torchvision>=0.18.0",
    "wandb>=0.17.0",  
    "numba>=0.59.0",
    "cloudpickle>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "twine>=4.0.2",
]
doc = [
    "docutils>=0.8,!=0.18.*,!=0.19.*",
    "sphinx>=4",
    "sphinx-book-theme>=1.0.0",
    "myst-nb",
    "sphinxcontrib-bibtex>=1.0.0",
    "sphinx-autodoc-typehints",
    "sphinxext-opengraph",
    # For notebooks
    "ipykernel",
    "ipython",
    "jupyterlab>=4.0.0",
    "sphinx-copybutton",
    "pandas",
]
test = [
    "pytest",
    "coverage",
]
cuda = [
    "nvidia-cublas-cu12>=12.1.3",
    "nvidia-cudnn-cu12>=8.9.2",
    "nvidia-cuda-runtime-cu12>=12.1.105",
    "triton>=2.3.0",
]

[tool.coverage.run]
source = ["nicheformer"]
omit = [
    "**/test_*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",  # allow using test files with same name
]

[tool.ruff]
line-length = 120
src = ["src"]
extend-include = ["*.ipynb"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "F",  # Errors detected by Pyflakes
    "E",  # Error detected by Pycodestyle
    "W",  # Warning detected by Pycodestyle
    "I",  # isort
    "D",  # pydocstyle
    "B",  # flake8-bugbear
    "TID",  # flake8-tidy-imports
    "C4",  # flake8-comprehensions
    "BLE",  # flake8-blind-except
    "UP",  # pyupgrade
    "RUF100",  # Report unused noqa directives
]
ignore = [
    # line too long -> we accept long comment lines; formatter gets rid of long code lines
    "E501",
    # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
    "E731",
    # allow I, O, l as variable names -> I is the identity matrix
    "E741",
    # Missing docstring in public package
    "D104",
    # Missing docstring in public module
    "D100",
    # Missing docstring in __init__
    "D107",
    # Errors from function calls in argument defaults. These are fine when the result is immutable.
    "B008",
    # __magic__ methods are are often self-explanatory, allow missing docstrings
    "D105",
    # first line should end with a period [Bug: doesn't work with single-line docstrings]
    "D400",
    # First line should be in imperative mood; try rephrasing
    "D401",
    ## Disable one in each pair of mutually incompatible rules
    # We don’t want a blank line before a class docstring
    "D203",
    # We want docstrings to start immediately after the opening triple quote
    "D213",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["I"]
"tests/*" = ["D"]
"*/__init__.py" = ["F401"]

[tool.cruft]
skip = [
    "tests",
    "src/**/__init__.py",
    "src/**/basic.py",
    "docs/api.md",
    "docs/changelog.md",
    "docs/references.bib",
    "docs/references.md",
    "docs/notebooks/example.ipynb",
]
